+++ src/main/java/sphabucks/payments/cards/model/Card.java
@@ -19,9 +19,9 @@ public class Card {
     @Column(nullable = false)
     private String name;    // 카드 이름
     @Column(updatable = false, nullable = false)
-    private Integer defaultMoney;   // 카드 초기금액
+    private Long defaultMoney;   // 카드 초기금액
     private String number;  // 카드 번호
     private String pin; // 핀번호(8자리)
     private String image;   // 카드 이미지 경로
-    private Integer money;  // 현재 잔액
+    private Long money;  // 현재 잔액
 }
