+++ src/main/java/sphabucks/products/repository/IBigCategoryRepository.java
@@ -3,6 +3,6 @@ package sphabucks.products.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sphabucks.products.model.BigCategory;
 
-public interface IBigCategoryRepository extends JpaRepository<BigCategory, Integer> {
+public interface IBigCategoryRepository extends JpaRepository<BigCategory, Long> {
 
 }
