+++ src/main/java/sphabucks/products/service/IProductCategoryListService.java
@@ -11,6 +11,6 @@ public interface IProductCategoryListService {
 
     void addProductCategoryList(RequestProductCategoryList requestProductCategoryList);
     List<ProductCategoryList> getByProductId(Long productId);
-    List<ResponseProductSummary> getProductsBySmallCategoryId(Integer smallCategoryId);
+    List<ResponseProductSummary> getProductsBySmallCategoryId(Long smallCategoryId);
     List<ResponseProductList> getOtherProductByProductId(Long productId); // 다른 고객이 함께 본 상품 api(소분류 카테고리, 동일 이벤트 검색)
 }
