+++ src/main/java/sphabucks/products/service/IProductService.java
@@ -15,7 +15,7 @@ public interface IProductService {
     List<Product> getAll();
 
     // 베스트 상품 조회 메서드 (대분류 카테고리별 조회)
-    List<ResponseProduct> getBestBigCategory(Integer bigCategoryId);
+    List<ResponseProduct> getBestBigCategory(Long bigCategoryId);
 
     // 상품 검색 메서드 (키워드 검색)
     List<ResponseSearchProduct> searchProductKeyword(String keyword, Pageable pageable);
